Add files via upload
This is a script that I have made to download workshop files from your workshop.txt file. This will automatically download the files for any server admins of QLDS. I have found that using your steamcmd and logging in with your steam account works far better with less issues than anonymous, however feel free to run as anonymous if you so choose. I guarantee nothing and this is as-is. I have taken bits and pieces from other scripts I've found on here to customize my own. There may be updates in the future to this.

Change the two variables to match your install location and your steamcmd.sh install location.
Also change steamID to your steamID you use to run your server. It may ask you for your 2FA (steam guard) if you have that setup.

I wanted to try to make this into a plugin because minqlx's getmap.py does not work for me. I haven't been able to get it to work. I have converted this bash code into python so I may continue to try working on that as well. Ideally I wanted to have a similar function where you type in the workshop ID and it'll download it live on the server instead of doing it through a bash script.